refactor: Make backend orchestrator, parser, and  plugins PEP8 (#879)
