+++ code/backend/batch/utilities/helpers/orchestrator_helper.py
@@ -1,8 +1,8 @@
 from typing import List
 
-from ..orchestrator.OrchestrationStrategy import OrchestrationStrategy
+from ..orchestrator.orchestration_strategy import OrchestrationStrategy
 from ..orchestrator import OrchestrationSettings
-from ..orchestrator.Strategies import get_orchestrator
+from ..orchestrator.strategies import get_orchestrator
 
 __all__ = ["OrchestrationStrategy"]
