+++ code/backend/batch/utilities/orchestrator/semantic_kernel.py
@@ -9,9 +9,9 @@ from semantic_kernel.contents.finish_reason import FinishReason
 
 from ..common.Answer import Answer
 from ..helpers.llm_helper import LLMHelper
-from ..plugins.ChatPlugin import ChatPlugin
-from ..plugins.PostAnsweringPlugin import PostAnsweringPlugin
-from .OrchestratorBase import OrchestratorBase
+from ..plugins.chat_plugin import ChatPlugin
+from ..plugins.post_answering_plugin import PostAnsweringPlugin
+from .orchestrator_base import OrchestratorBase
 
 logger = logging.getLogger(__name__)
