+++ code/backend/batch/utilities/plugins/chat_plugin.py
