+++ code/tests/utilities/helpers/processors/test_Integrated_vectorization_embedder.py
