+++ code/tests/utilities/helpers/test_azure_search_helper.py
