+++ code/tests/utilities/integrated_vectorization/test_azure_search_indexer.py
